Exam test문제 업데이트

+++ Exam/src/test1/Test01.java
@@ -9,7 +9,7 @@ public class Test01 {
 	public static void main(String[] args) {
 
 		char var1 = 'A';
-		byte var2 = 100;
+		int var2 = 100;
 		long var3 = 2147483648L;
 		float var4 = 3.14f;
 		double var5 = 3.14159;
